pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.115.14",
    "httpx>=0.28.1",
    "langchain-community>=0.3.27",
    "langchain>=0.3.26",
    "llama-cloud-services>=0.6.41",
    "openai>=1.93.0",
    "pydantic-settings>=2.10.1",
    "qdrant-client>=1.14.3",
    "langchain-openai>=0.3.27",
    "langchain-qdrant>=0.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
]
